+++ packages/widget/src/types/index.d.ts
@@ -9,6 +9,7 @@ export interface BridgeableToken {
   color: string
   priorityRank: number
   routeSymbol: string
+  imgUrl: string
 }
 
 export interface CustomThemeVariables {
@@ -60,6 +61,7 @@ export interface Chain {
   explorerUrl: string
   explorerName: string
   blockTime: number
+  imgUrl: string
 }
 
 export declare function Bridge(props: BridgeProps): JSX.Element